+++ tests/reuse1.c
@@ -100,7 +100,7 @@ main()
   washere = 0;
   assert(pthread_create(&t, &attr, func, NULL) == 0);
   assert(pthread_join(t, &result) == 0);;
-  assert(result == 0);
+  assert((int)(size_t)result == 0);
   assert(washere == 1);
   last_t = t;
 
@@ -109,7 +109,7 @@ main()
       washere = 0;
       assert(pthread_create(&t, &attr, func, (void *) i) == 0);
       pthread_join(t, &result);
-      assert((int) result == i);
+      assert((int)(size_t) result == i);
       assert(washere == 1);
       /* thread IDs should be unique */
       assert(!pthread_equal(t, last_t));
